requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
groq
python-dotenv